chore: update formatter for V3 Module and DeleteUsecase
